Create Multithread Fibonacci Sequence
You are tasked with implementing a multithreaded C program that performs the following two tasks using threads:
1.Fibonacci Sequence Generation
The first thread should compute the Fibonacci sequence up to the n-th term (inclusive), where n is an integer input from the user. The result should be returned to the main thread using dynamic memory allocation.

2.Fibonacci Value Search
The second thread should take the computed Fibonacci sequence and search for multiple terms specified by the user. For each search index, it should return the corresponding Fibonacci value if the index is valid, or -1 if the index is invalid.